# backend/api-doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/rent/returning": {
      "post": {
        "tags": [
          "rent-controller"
        ],
        "operationId": "returningLoan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonRentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rent/return": {
      "post": {
        "tags": [
          "rent-controller"
        ],
        "operationId": "returnLoan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonRentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rent/reject": {
      "post": {
        "tags": [
          "rent-controller"
        ],
        "operationId": "rejectLoan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonRentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rent/approve": {
      "post": {
        "tags": [
          "rent-controller"
        ],
        "operationId": "approveLoan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonRentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rent/": {
      "post": {
        "tags": [
          "rent-controller"
        ],
        "summary": "rent a book",
        "operationId": "rentBook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonRentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rent/getRentsByUserId": {
      "get": {
        "tags": [
          "rent-controller"
        ],
        "operationId": "getRentsByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonListRentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rent/getLoansByUserIdAndStatus": {
      "get": {
        "tags": [
          "rent-controller"
        ],
        "operationId": "getLoansByUserIdAndStatus",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonListLoanBookResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rent/get-daily-hourly-loan-data": {
      "get": {
        "tags": [
          "rent-controller"
        ],
        "summary": "get daily hourly loan data",
        "operationId": "getHourlyLoanData",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonObject"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoanStateRequest": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "integer",
            "format": "int32"
          },
          "bookCopyId": {
            "type": "integer",
            "format": "int32"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "BookCopyResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "barcode": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "loanInfo": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoanInfo"
            }
          }
        }
      },
      "LoanInfo": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "integer",
            "format": "int32"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "RentResponse": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "integer",
            "format": "int32"
          },
          "userName": {
            "type": "string"
          },
          "bookCopyResponse": {
            "$ref": "#/components/schemas/BookCopyResponse"
          },
          "memberId": {
            "type": "integer",
            "format": "int32"
          },
          "maxBook": {
            "type": "integer",
            "format": "int32"
          },
          "memFee": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "borrowAt": {
            "type": "string",
            "format": "date-time"
          },
          "returnAt": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResponseCommonRentResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/RentResponse"
          }
        }
      },
      "RentRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "borrowAt": {
            "type": "string"
          },
          "returnAt": {
            "type": "string"
          },
          "barcode": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "ResponseCommonListRentResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentResponse"
            }
          }
        }
      },
      "LoanBookResponse": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "borrowAt": {
            "type": "string"
          },
          "returnAt": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "bookCopyResponse": {
            "$ref": "#/components/schemas/BookCopyResponse"
          }
        }
      },
      "ResponseCommonListLoanBookResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoanBookResponse"
            }
          }
        }
      },
      "ResponseCommonObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
    }
  }
}